+++ segmentation_robustness_framework/__init__.py
@@ -4,9 +4,9 @@ from .datasets import ade20k, cityscapes, stanford_background, voc
 from .models import deeplab, fcn
 from .robustness_evaluation import RobustnessEvaluation
 from .utils import image_preprocessing, metrics
-from .engine import model_loader
-from .engine import dataset_loader
-from .engine import attack_loader
+from .loaders import model_loader
+from .loaders import dataset_loader
+from .loaders import attack_loader
 
 __all__ = [
     "fcn",
